✨ refactor(grid): move roundAndTruncatePrice to util package and export it for sharing among modules
