+++ pkg/strategy/grid2/grid_test.go
@@ -215,40 +215,6 @@ func Test_calculateArithmeticPins(t *testing.T) {
 	}
 }
 
-func Test_filterPrice1(t *testing.T) {
-	type args struct {
-		p    fixedpoint.Value
-		prec int
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{
-			name: "basic",
-			args: args{p: number("31.2222"), prec: 3},
-			want: "31.222",
-		},
-		{
-			name: "roundup",
-			args: args{p: number("31.22295"), prec: 3},
-			want: "31.223",
-		},
-		{
-			name: "roundup2",
-			args: args{p: number("31.22290"), prec: 3},
-			want: "31.222",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			rst := roundAndTruncatePrice(tt.args.p, tt.args.prec)
-			assert.Equalf(t, tt.want, rst.String(), "roundAndTruncatePrice(%v, %v)", tt.args.p, tt.args.prec)
-		})
-	}
-}
-
 func Test_removeDuplicatedPins(t *testing.T) {
 	pins := []Pin{
 		Pin(number("31.222")),
